+++ examples/01_plotting/plot_demo_glass_brain_extensive.py
@@ -2,11 +2,11 @@
 Glass brain plotting in nilearn (all options)
 =============================================
 
-First part of this example goes through different options of the
-:func:`nilearn.plotting.plot_glass_brain` function (including plotting
+The first part of this example goes through different options of the
+:func:`~nilearn.plotting.plot_glass_brain` function (including plotting
 negative values).
 
-Second part, goes through same options but selected of the same glass brain
+The second part goes through same options but selected of the same glass brain
 function but plotting is seen with contours.
 
 See :ref:`plotting` for more plotting functionalities and
@@ -41,173 +41,262 @@ stat_img
 ###############################################################################
 # Demo glass brain plotting
 # --------------------------
+#
+# By default, :func:`~nilearn.plotting.plot_glass_brain` uses a display mode
+# called 'ortho' which results in three projections. It is equivalent to
+# specify ``display_mode='ortho'`` in
+# :func:`~nilearn.plotting.plot_glass_brain`. Note that depending on the
+# value of ``display_mode``, different display objects are returned. Here,
+# a :class:`~nilearn.plotting.displays.OrthoProjector` is returned.
+
 from nilearn import plotting
+from nilearn.plotting import plot_glass_brain
 
 # Whole brain sagittal cuts and map is thresholded at 3
-plotting.plot_glass_brain(stat_img, threshold=3)
+plot_glass_brain(stat_img, threshold=3)
 
 
 ###############################################################################
-# With a colorbar
-plotting.plot_glass_brain(stat_img, threshold=3, colorbar=True)
+# The same figure, with a colorbar, can be produced by
+# setting ``colorbar=True``.
+
+plot_glass_brain(stat_img, threshold=3, colorbar=True)
 
 
 ###############################################################################
-# Black background, and only the (x, z) cuts
-plotting.plot_glass_brain(stat_img, title='plot_glass_brain',
-                          black_bg=True, display_mode='xz', threshold=3)
+# Here, we show how to set a black background, and we only view sagittal and
+# axial projections by setting ``display_mode='xz'``, which returns a
+# :class:`~nilearn.plotting.displays.XZProjector`.
+
+plot_glass_brain(
+    stat_img, title='plot_glass_brain', black_bg=True,
+    display_mode='xz', threshold=3
+)
 
 
 ###############################################################################
-# Plotting the sign of the activation with plot_abs to False
-plotting.plot_glass_brain(stat_img, threshold=0, colorbar=True,
-                          plot_abs=False)
+# We can also plot the sign of the activation by setting ``plot_abs=False``.
+# Additionally, we only visualize coronal and axial projections by setting
+# ``display_mode='yz'`` which returns a
+# :class:`~nilearn.plotting.displays.YZProjector`.
+
+plot_glass_brain(
+    stat_img, threshold=0, colorbar=True, plot_abs=False, display_mode='yz'
+)
 
 
 ###############################################################################
-# The sign of the activation and a colorbar
-plotting.plot_glass_brain(stat_img, threshold=3,
-                          colorbar=True, plot_abs=False)
+# Setting ``plot_abs=True`` and ``display_mode='yx'`` (returns a
+# :class:`~nilearn.plotting.displays.YXProjector`).
+
+plot_glass_brain(
+    stat_img, threshold=3, colorbar=True, plot_abs=False, display_mode='yx'
+)
 
 
 ###############################################################################
 # Different projections for the left and right hemispheres
 # ---------------------------------------------------------
 #
-# Hemispheric sagittal cuts
-plotting.plot_glass_brain(stat_img,
-                          title='plot_glass_brain with display_mode="lzr"',
-                          black_bg=True, display_mode='lzr', threshold=3)
+# In the previous section we saw a few projection modes, which are controlled
+# by setting the argument ``display_mode`` of
+# :func:`~nilearn.plotting.plot_glass_brain`. In this section, we will show
+# some additional possibilities. For example, setting ``display_mode='lzr'``
+# enables an hemispheric sagittal view. The display object returned is then a
+# :class:`~nilearn.plotting.displays.LZRProjector`.
+
+plot_glass_brain(
+    stat_img, title='plot_glass_brain with display_mode="lzr"',
+    black_bg=True, display_mode='lzr', threshold=3
+)
+
 
 ###############################################################################
-plotting.plot_glass_brain(stat_img, threshold=0, colorbar=True,
-                          title='plot_glass_brain with display_mode="lyrz"',
-                          plot_abs=False, display_mode='lyrz')
+# ``display_mode='lyrz'`` returns a
+# :class:`~nilearn.plotting.displays.LYRZProjector` object.
+
+plot_glass_brain(
+    stat_img, threshold=0, colorbar=True,
+    title='plot_glass_brain with display_mode="lyrz"',
+    plot_abs=False, display_mode='lyrz'
+)
+
+
+###############################################################################
+# If you are only interested in single projections, you can set
+# ``display_mode`` to 'x' (returns a
+# :class:`~nilearn.plotting.displays.XProjector`), 'y' (returns a
+# :class:`~nilearn.plotting.displays.YProjector`), 'z' (returns a
+# :class:`~nilearn.plotting.displays.ZProjector`), 'l' (returns a
+# :class:`~nilearn.plotting.displays.LProjector`), or 'r' (returns a
+# :class:`~nilearn.plotting.displays.RProjector`).
+
+plot_glass_brain(
+    stat_img, threshold=0, colorbar=True, title='display_mode="x"',
+    plot_abs=False, display_mode='x'
+)
+
 
 ###############################################################################
 # Demo glass brain plotting with contours and with fillings
 # ---------------------------------------------------------
-# To plot maps with contours, we call the plotting function into variable from
-# which we can use specific display features which are inherited automatically.
-# In this case, we focus on using add_contours
-# First, we initialize the plotting function into "display" and first
-# argument set to None since we want an empty glass brain to plotting the
-# statistical maps with "add_contours"
-display = plotting.plot_glass_brain(None)
+#
+# The display objects returned by :func:`~nilearn.plotting.plot_glass_brain`
+# all inherit from the :class:`~nilearn.plotting.displays.OrthoProjector`
+# and enable further customisation of the figures.
+#
+# In this example, we focus on using methods
+# :meth:`~nilearn.plotting.displays.OrthoProjector.add_contours` and
+# :meth:`~nilearn.plotting.displays.OrthoProjector.title`. First, we
+# save the display object (here a
+# :class:`~nilearn.plotting.displays.LZRYProjector`) into a variable named
+# ``display``. Note that we set the first argument to ``None`` since we
+# want an empty glass brain to plot the statistical maps with
+# :meth:`~nilearn.plotting.displays.OrthoProjector.add_contours`.
+
+display = plot_glass_brain(None, display_mode='lzry')
 # Here, we project statistical maps
 display.add_contours(stat_img)
-# and a title
+# and add a title
 display.title('"stat_img" on glass brain without threshold')
 
+
 ###############################################################################
-# Plotting with `filled=True` implies contours with fillings. Here, we are not
-# specifying levels
-display = plotting.plot_glass_brain(None)
+# We can fill the contours by setting ``filled=True``. Note that we are not
+# specifying levels here
+
+display = plot_glass_brain(None, display_mode='lzry')
 # Here, we project statistical maps with filled=True
 display.add_contours(stat_img, filled=True)
-# and a title
+# and add a title
 display.title('Same map but with fillings in the contours')
 
-###############################################################################
-# Here, we input specific level (cut-off) in the statistical map. In other way,
-# we are thresholding our statistical map
 
-# Here, we set the threshold using parameter called `levels` with value given
-# in a list and choosing color to Red.
-display = plotting.plot_glass_brain(None)
+###############################################################################
+# Here, we input a specific level (cut-off) in the statistical map.
+# In other words, we are thresholding our statistical map.
+#
+# We set the threshold using a parameter of method
+# :meth:`~nilearn.plotting.displays.OrthoProjector.add_contours` called
+# ``levels`` which value is given as a list and we choose the color to be red.
+
+display = plot_glass_brain(None, display_mode='lzry')
 display.add_contours(stat_img, levels=[3.], colors='r')
 display.title('"stat_img" on glass brain with threshold')
 
+
 ###############################################################################
-# Plotting with same demonstration but includes now filled=True
-display = plotting.plot_glass_brain(None)
+# Plotting with same demonstration but fill the contours (by setting
+# ``filled=True``).
+
+display = plot_glass_brain(None, display_mode='lzry')
 display.add_contours(stat_img, filled=True, levels=[3.], colors='r')
 display.title('Same demonstration but using fillings inside contours')
 
+
 ##############################################################################
-# Plotting with black background, `black_bg` should be set with
-# `plot_glass_brain`
+# Plotting with black background, ``black_bg`` should be set to ``True``
+# through :func:`~nilearn.plotting.plot_glass_brain`.
 
 # We can set black background using black_bg=True
-display = plotting.plot_glass_brain(None, black_bg=True)
+display = plot_glass_brain(None, black_bg=True)
 display.add_contours(stat_img, levels=[3.], colors='g')
 display.title('"stat_img" on glass brain with black background')
 
+
 ##############################################################################
-# Black background plotting with filled in contours
-display = plotting.plot_glass_brain(None, black_bg=True)
+# Black background plotting with filled in contours.
+
+display = plot_glass_brain(None, black_bg=True)
 display.add_contours(stat_img, filled=True, levels=[3.], colors='g')
 display.title('Glass brain with black background and filled in contours')
+
 
 ##############################################################################
 # Display contour projections in both hemispheres
 # -------------------------------------------------
-# Key argument to vary here is `display_mode` for hemispheric plotting
+#
+# The key argument to vary here is ``display_mode`` for hemispheric plotting.
+# Here, we set ``display_mode='lr'`` for both hemispheric plots. Note that a
+# :class:`~nilearn.plotting.displays.LRProjector` is returned.
 
-# Now, display_mode is chosen as 'lr' for both hemispheric plots
-display = plotting.plot_glass_brain(None, display_mode='lr')
+display = plot_glass_brain(None, display_mode='lr')
 display.add_contours(stat_img, levels=[3.], colors='r')
 display.title('"stat_img" on glass brain only "l" "r" hemispheres')
 
+
 ##############################################################################
-# Filled contours in both hemispheric plotting, just by adding filled=True
-display = plotting.plot_glass_brain(None, display_mode='lr')
+# Filled contours in both hemispheric plotting, by adding ``filled=True``.
+
+display = plot_glass_brain(None, display_mode='lr')
 display.add_contours(stat_img, filled=True, levels=[3.], colors='r')
 display.title('Filled contours on glass brain only "l" "r" hemispheres')
 
-##############################################################################
-# With positive and negative sign of activations with `plot_abs` in
-# `plot_glass_brain`
 
-# By default parameter `plot_abs` is True and sign of activations can be
-# displayed by changing `plot_abs` to False
-display = plotting.plot_glass_brain(None, plot_abs=False, display_mode='lzry')
+##############################################################################
+# With positive and negative signs of activations with ``plot_abs`` in
+# :func:`~nilearn.plotting.plot_glass_brain`.
+#
+# By default parameter ``plot_abs`` is ``True`` and sign of activations
+# can be displayed by changing ``plot_abs`` to ``False``. Note that we also
+# specify ``display_mode='lyr'`` which returns a
+# :class:`~nilearn.plotting.displays.LYRProjector` display object.
+
+display = plot_glass_brain(None, plot_abs=False, display_mode='lyr')
 display.add_contours(stat_img)
 display.title("Contours with both sign of activations without threshold")
 
+
 ##############################################################################
-# Now, adding just filled=True to get positive and negative sign activations
-# with fillings in the contours
-display = plotting.plot_glass_brain(None, plot_abs=False, display_mode='lzry')
+# Now, adding ``filled=True`` to get positive and negative sign activations
+# with fillings in the contours.
+
+display = plot_glass_brain(None, plot_abs=False, display_mode='lyr')
 display.add_contours(stat_img, filled=True)
-display.title("Filled contours with both sign of activations without threshold")
+display.title(
+    "Filled contours with both sign of activations without threshold"
+)
 
 
 ##############################################################################
 # Displaying both signs (positive and negative) of activations with threshold
-# meaning thresholding by adding an argument `levels` in add_contours.
+# meaning thresholding by adding an argument ``levels`` in method
+# :meth:`~nilearn.plotting.displays.OrthoProjector.add_contours`.
+#
+# We give two values through the argument ``levels`` which corresponds to the
+# thresholds of the contour we want to draw: One is positive and the other one
+# is negative. We give a list of ``colors`` as argument to associate a
+# different color to each contour. Additionally, we also choose to plot
+# contours with thick line widths. For ``linewidths``, one value would be
+# enough so that same value is used for both contours.
 
 import numpy as np
-display = plotting.plot_glass_brain(None, plot_abs=False, display_mode='lzry')
-
-# In add_contours,
-# we give two values through the argument `levels` which corresponds to the
-# thresholds of the contour we want to draw: One is positive and the other one
-# is negative. We give a list of `colors` as argument to associate a different
-# color to each contour. Additionally, we also choose to plot contours with
-# thick line widths, For linewidths one value would be enough so that same
-# value is used for both contours.
-display.add_contours(stat_img, levels=[-2.8, 3.], colors=['b', 'r'],
-                     linewidths=4.)
+display = plot_glass_brain(None, plot_abs=False, display_mode='lzry')
+display.add_contours(
+    stat_img, levels=[-2.8, 3.], colors=['b', 'r'], linewidths=4.
+)
 display.title('Contours with sign of activations with threshold')
 
+
 ##############################################################################
-# Same display demonstration as above but just adding filled=True to get
+# Same display demonstration as above but adding ``filled=True`` to get
 # fillings inside the contours.
+#
+# Unlike in previous plot, here we specify each sign at a time. We call
+# negative values display first followed by positive values display.
+#
+# First, we fetch our display object with same parameters used as above.
+# Then, we plot negative sign of activation with levels given as negative
+# activation value in a list. Upper bound should be kept to -infinity.
+# Next, using the same display object, we plot positive sign of activation.
 
-# Unlike in previous plot, here we specify each sign at a time. We call negative
-# values display first followed by positive values display.
-
-# First, we fetch our display object with same parameters used as above
-display = plotting.plot_glass_brain(None, plot_abs=False, display_mode='lzry')
-
-# Second, we plot negative sign of activation with levels given as negative
-# activation value in a list. Upper bound should be kept to -infinity
-display.add_contours(stat_img, filled=True, levels=[-np.inf, -2.8],
-                     colors='b')
-# Next, within same plotting object we plot positive sign of activation
-display.add_contours(stat_img, filled=True, levels=[3.], colors='r')
+display = plot_glass_brain(None, plot_abs=False, display_mode='lzry')
+display.add_contours(
+    stat_img, filled=True, levels=[-np.inf, -2.8], colors='b'
+)
+display.add_contours(
+    stat_img, filled=True, levels=[3.], colors='r'
+)
 display.title('Now same plotting but with filled contours')
-
 # Finally, displaying them
 plotting.show()